server/requirements/base.txt
Django~=4.2.6
djangorestframework==3.14.0
drf-yasg
django-environ
